final changes

+++ src/contracts/IERC20.sol
@@ -3,7 +3,13 @@ pragma solidity ^0.5.16;
 contract IERC20 {
     function totalSupply() external view returns (uint256);
 
-    function permit(address owner, address spender, uint value, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
+    function permit(  address owner,
+        address spender,
+        uint256 value,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s) external;
 
     function balanceOf(address account) external view returns (uint256);
